cognita-vector-service/requirements.txt
sentence-transformers
faiss-cpu
numpy
flask
